relocate cuda_setup module ito device_setup/cuda
