+++ bitsandbytes/device_setup/cuda/main.py
@@ -142,7 +142,7 @@ class CUDASetup:
         self.binary_name = binary_name
         self.manual_override()
 
-        package_dir = Path(__file__).parent.parent
+        package_dir = Path(__file__).parent.parent.parent
         binary_path = package_dir / self.binary_name
 
         try:
@@ -278,7 +278,7 @@ def determine_cuda_runtime_lib_path() -> Union[Path, None]:
             1. active conda env
             2. LD_LIBRARY_PATH
             3. any other env vars, while ignoring those that
-                - are known to be unrelated (see `bnb.cuda_setup.env_vars.to_be_ignored`)
+                - are known to be unrelated (see `bnb.device_setup.cuda.env_vars.to_be_ignored`)
                 - don't contain the path separator `/`
 
         If multiple libraries are found in part 3, we optimistically try one,
